Implement the prime-sieve example, adds std/lazy.
The sieve is implemented as a Sieve of Eratosthenes as per SICP's
implementation. The laziness system is similarly inspired, although it
uses the (not yet implemented) std/stm module to provide mutability, as
set! is not possible in OftLisp.

Additionally, lazy streams will provide as many as possible of the
methods typically defined over lists using the class system. The class
system is mostly designed, but still awaiting the macro system before
implementation can begin.